requirements.txt
discord.py
yt-dlp
youtube-search-python
requests